Cleaned up deep dive visualization code, fixed up visualizations a bit, updated requirements.txt and setup.py with new required packages.
