+++ requirements.txt
@@ -9,3 +9,7 @@ fedex-generator>=1.0.5
 cluster-explorer >= 1.0.2
 external-explainers>=1.0.0
 pytest>=6.2.4
+ipywidgets>=8.1.0
+ipycytoscape>=1.3.0
+together>=1.4.6
+openai>=1.66.5
